feat(sidebar): open the switcher straight from the collapsed rail
Clicking the avatar expands the sidebar and hands over the menu, so the
rail costs a click instead of hiding the control behind one.
