+++ __tests__/components/organization-switcher.test.tsx
@@ -37,6 +37,7 @@ vi.mock('@/components/ui/sidebar', () => ({
     isResponsiveOverride: false,
     isMobile: false,
     setOpenMobile: vi.fn(),
+    expand: vi.fn(),
   })
 }))
